+++ src/plugins/core/registerCorePlugin.ts
@@ -26,6 +26,7 @@ function initOnRequestHook(api: models.HttpyacHooksApi) {
   api.hooks.onRequest.addHook('attachDefaultHeaders', request.attachDefaultHeaders);
   api.hooks.onRequest.addHook('setEnvRequestOptions', request.setEnvRequestOptions);
   api.hooks.onRequest.addHook('resolveRequestBody', request.resolveRequestBody);
+  api.hooks.onRequest.addHook('setDefaultHttpyacHeaders', request.setDefaultHttpyacHeaders);
   api.hooks.onRequest.addHook('requestVariableReplacer', request.requestVariableReplacer);
   api.hooks.onRequest.addHook('transformRequestBody', request.transformRequestBodyToBuffer);
   api.hooks.onRequest.addHook('transfromMultilineFormUrlEncoded', request.transfromMultilineFormUrlEncoded);