+++ src/plugins/core/request/index.ts
@@ -4,6 +4,7 @@ export * from './excludeProxyInterceptor';
 export * from './isTrustedInterceptor';
 export * from './requestVariableReplacer';
 export * from './resolveRequestBody';
+export * from './setDefaultHttpyacHeaders';
 export * from './setEnvRequestOptions';
 export * from './transformRequestBodyToBuffer';
 export * from './transfromMultilineFormUrlEncoded';