+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [unreleased]
 ### Fix
 - variables in gql Query body are replaced (AnWeber/vscode-httpyac#303)
+- allow defaultHeaders to overwrite accept or user-agent header
 
 ## [6.14.0] ( 2024-06-01)
 ### Features